// brokerchain-1.0.0/src/library/buffer/buffer-slice-batch/type.ts
export interface Input {
    hex: string;
    byte_start: number;
    byte_length_list: number[];
}

export interface OutputEmpty {
    sliced_hex_list: string[];
}

export interface OutputOk {
    sliced_hex_list: string[];
}

export interface OutputFail {}

export interface Callback<R> {
    empty: (output: OutputEmpty) => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;